adding firstchart.py
So far this is a function to read the data and save it to a smaller
file. It worked, but I got warning:

PerformanceWarning: your performance may suffer as PyTables will pickle object types that it cannot map directly to c-types

I haven't looked in to what that means.

Otherwise my next steps will be to read the new file, get the series I
want, and then plot them.